refactor(normalizer): Extract own package
Normalizer do not need a dependency to ckeditor 5 but have to be used in the CMS context. Therefore, the normalizer must not have a ckeditor 5 package dependency.
