+++ app/package.json
@@ -41,6 +41,7 @@
     "@coremedia/ckeditor5-dataprocessor-support": "6.1.1-rc.1",
     "@coremedia/ckeditor5-studio-essentials": "6.1.1-rc.1",
     "@coremedia/ckeditor5-symbol-on-paste-mapper": "6.1.1-rc.1",
+    "@coremedia/ckeditor5-richtext-normalizer": "6.1.1-rc.1",
     "@coremedia/service-agent": "^1.1.0",
     "xml-formatter": "^2.6.1"
   },
