+++ packages/ckeditor5-richtext-normalizer/__tests__/DataDiffer.test.ts
