+++ packages/ckeditor5-richtext-normalizer/__tests__/NormalizedData.test.ts
