+++ packages/ckeditor5-richtext-normalizer/babel.config.js
@@ -1,0 +1,3 @@
+const sharedBabelConf = require('@coremedia-internal/ckeditor5-babel-config');
+
+module.exports = sharedBabelConf;
