+++ packages/ckeditor5-richtext-normalizer/jest.config.js
@@ -1,0 +1,3 @@
+const jestConfig = require("@coremedia-internal/ckeditor5-jest-test-helpers/shared-jest.config.js");
+
+module.exports = { ...jestConfig};
