+++ packages/ckeditor5-richtext-normalizer/src/DataDiffer.ts
