+++ packages/ckeditor5-richtext-normalizer/src/NormalizedData.ts
