+++ packages/ckeditor5-richtext-normalizer/src/Normalizers.ts
